requirements.txt
pytest
pytest-cov
pytest-profiling
numpy==1.19.4
torch
flask
websocket_client